Messages and message handlers for Chord
Note: Also incorporated logging

+++ utils.go
@@ -14,11 +14,11 @@ func bisectLeftNode(items []Node, searchKey uint64) int {
 	return sort.Search(len(items), func(i int) bool { return items[i].id >= searchKey })
 }
 
-func calculateId(key string, m uint) (uint64) {
+func calculateId(key string, m int) (uint64) {
 	h := sha1.New()
 	h.Write([]byte(key))
 	hashedKey := binary.BigEndian.Uint64(h.Sum(nil))
-	return hashedKey % (1 << m)
+	return hashedKey % (1 << uint(m))
 }
 
 func assertIdsUnique(records []Record) bool {
